+++ src/components/roof_selection/SurfaceNumber.vue
@@ -1,20 +1,24 @@
 <script setup lang="ts">
 import SurfaceChartDonut from '@/components/chart/SurfaceChartDonut.vue'
 import UiSurfaceNumber from '@/components/roof_selection/UiSurfaceNumber.vue'
-import { useRoofDataStore } from '@/stores/roofData'
-const roofDataStore = useRoofDataStore()
+import { ref } from 'vue'
+import { getDataBuilding } from '@/model/roof.model'
+import { useRoofsStore } from '@/stores/roof'
+
+const roofsStore = useRoofsStore()
+const buildingData = ref(getDataBuilding())
+
+roofsStore.$subscribe(async () => {
+  buildingData.value = getDataBuilding()
+})
 </script>
 
 <template>
-  <template v-if="roofDataStore.buildingData !== null">
-    <div class="flex flex-row items-center justify-center">
-      <SurfaceChartDonut></SurfaceChartDonut>
-    </div>
-  </template>
-  <template v-if="roofDataStore.buildingData !== null">
-    <UiSurfaceNumber
-      :totalArea="roofDataStore.buildingData.total"
-      :favorableArea="roofDataStore.buildingData.favorable"
-    ></UiSurfaceNumber>
-  </template>
+  <div class="flex flex-row items-center justify-center">
+    <SurfaceChartDonut :buildingData="buildingData"></SurfaceChartDonut>
+  </div>
+  <UiSurfaceNumber
+    :totalArea="buildingData.total"
+    :favorableArea="buildingData.favorable"
+  ></UiSurfaceNumber>
 </template>
