+++ src/model/roof.model.ts
@@ -2,6 +2,7 @@ import type { GeoJSONFeature } from 'ol/format/GeoJSON'
 import GeoJSON from 'ol/format/GeoJSON'
 import type { Feature } from 'ol'
 import type { Geometry } from 'ol/geom'
+import { useRoofsStore } from '@/stores/roof'
 
 export type RoofSurfaceModel = {
   surface_id: string
@@ -59,4 +60,31 @@ export function mapRoofSurfaceModel(geojson: GeoJSONFeature): RoofSurfaceModel {
     azimuth: feature.getProperty('azimuth'),
     inclinaison: feature.getProperty('slope'),
   }
+}
+
+export function getDataBuilding() {
+  let all_area = 0
+  let all_area_favorable = 0
+  const all_values = [0, 0, 0, 0]
+  const roofsStore = useRoofsStore()
+  const buildingRoofsFeatures = roofsStore.buildingRoofsFeatures
+  buildingRoofsFeatures?.features.forEach((feature) => {
+    const dataRoof: RoofSurfaceModel = mapRoofSurfaceModel(feature)
+
+    all_area += dataRoof.total
+    all_area_favorable += dataRoof.favorable
+
+    for (let i = 0; i < 4; i++) {
+      all_values[i] += (dataRoof.values[i] * dataRoof.total) / 100
+    }
+  })
+  for (let i = 0; i < 4; i++) {
+    all_values[i] = (all_values[i] * 100) / all_area
+  }
+  const dataBuilding = {
+    total: roundResult(all_area),
+    favorable: roundResult(all_area_favorable),
+    values: all_values,
+  }
+  return dataBuilding
 }
